4F

+++ contest4/E.cpp
@@ -9,13 +9,13 @@ int main()
     cin >> n;
     for(int i = 0; i < n; i++)
     {
-        int w, h;
+        unsigned w, h;
         cin >> w >> h;
         if(h > bricks[w])
             bricks[w] = h;
     }
 
-    int maxH = 0;
+    long unsigned maxH = 0;
     for(auto&[w, h] : bricks)
     {
         maxH += h;
